extension/comic-dl/requirements.txt
requests==2.22.0
beautifulsoup4==4.9.3
Pillow==8.1.0
PyYAML==5.4.1
pysean==0.2.3
